# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Interior Design App API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CreateProjectInput {
  photoUri: string;
  roomType: string;
}

export interface Project {
  id: string;
  photoUri: string;
  roomType: string;
  aiRoomImageUrl: string;
  createdAt: string;
}

export interface UserProfileInput {
  stylePreference: string;
  colorPalette: string;
  budget: number;
  projectId: string;
}

export interface UserProfile {
  id: string;
  stylePreference: string;
  colorPalette: string;
  budget: number;
  projectId: string;
  createdAt: string;
}

export interface FurnitureItem {
  id: string;
  name: string;
  brand: string;
  price: number;
  imageUrl: string;
  category: string;
  description: string;
  inStock: boolean;
}

export interface AddCartItemInput {
  furnitureItemId: string;
  quantity: number;
}

export interface CartItem {
  id: string;
  furnitureItemId: string;
  quantity: number;
  furnitureItem: FurnitureItem;
}

export interface Order {
  id: string;
  status: string;
  totalAmount: number;
  itemCount: number;
  createdAt: string;
}

export interface SuccessResponse {
  success: boolean;
}

export type ListFurnitureItemsParams = {
projectId?: string;
};