# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Interior Design App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cart-items:
    get:
      operationId: listCartItems
      tags: [cartItems]
      summary: Get all cart items
      responses:
        "200":
          description: Cart items list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CartItem"
    post:
      operationId: addCartItem
      tags: [cartItems]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCartItemInput"
      responses:
        "201":
          description: Item added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartItem"
  /cart-items/{id}:
    delete:
      operationId: removeCartItem
      tags: [cartItems]
      summary: Remove item from cart
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Item removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    FurnitureItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        brand:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        category:
          type: string
        description:
          type: string
        inStock:
          type: boolean
      required:
        - id
        - name
        - brand
        - price
        - imageUrl
        - category
        - description
        - inStock
    AddCartItemInput:
      type: object
      properties:
        furnitureItemId:
          type: string
        quantity:
          type: number
      required:
        - furnitureItemId
        - quantity
    CartItem:
      type: object
      properties:
        id:
          type: string
        furnitureItemId:
          type: string
        quantity:
          type: number
        furnitureItem:
          $ref: "#/components/schemas/FurnitureItem"
      required:
        - id
        - furnitureItemId
        - quantity
        - furnitureItem
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success