# lib/api-zod/src/generated/types/addCartItemInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Interior Design App API
 * OpenAPI spec version: 0.1.0
 */

export interface AddCartItemInput {
  furnitureItemId: string;
  quantity: number;
}