# lib/api-zod/src/generated/types/cartItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Interior Design App API
 * OpenAPI spec version: 0.1.0
 */
import type { FurnitureItem } from './furnitureItem';

export interface CartItem {
  id: string;
  furnitureItemId: string;
  quantity: number;
  furnitureItem: FurnitureItem;
}