# lib/api-zod/src/generated/types/furnitureItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Interior Design App API
 * OpenAPI spec version: 0.1.0
 */

export interface FurnitureItem {
  id: string;
  name: string;
  brand: string;
  price: number;
  imageUrl: string;
  category: string;
  description: string;
  inStock: boolean;
}